// src/types/models/PriceResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PriceResponse = {
    id: number;
    product_id: number;
    weight: string;
    price: number;
    old_price?: number;
    quantity: number;
    is_active: boolean;
    is_deleted: boolean;
    promotional: boolean;
};